news
перенес логику в другой файл,
добавил троеточия,
преобразование даны в правильный формат,
